chore: create new component
